+++ .github/workflows/test.yml
@@ -44,4 +44,4 @@ jobs:
       - name: Pytest
         run: |
           . $HOME/.poetry/env
-          poetry run pytest -s -q --all_data
+          poetry run pytest -s -q --all_data tests/PUMS_Tests
