+++ tests/PUMS_Tests/test_PUMS_download.py
@@ -4,7 +4,7 @@ clean/collate data correctly"""
 
 
 import pandas as pd
-from tests.local_loader import LocalLoader
+from tests.PUMS_Tests.local_loader import LocalLoader
 
 local_loader = LocalLoader()
