+++ tests/PUMS_Tests/test_PUMS_fractions_demographics_aggregation.py
@@ -4,7 +4,7 @@ should all be close to one for each indicator. categories attribute is dictionar
 
 import pytest
 from tests.util import races
-from tests.local_loader import LocalLoader
+from tests.PUMS_Tests.local_loader import LocalLoader
 import numpy as np
 import itertools
